add module scope v1 for the app version
